+++ Source/NETworkManager.Models/Network/SNMPV3AuthenticationProvider.cs
@@ -1,0 +1,10 @@
+﻿namespace NETworkManager.Models.Network;
+
+/// <summary>
+/// Enum for the SNMP v3 authentication provider.
+/// </summary>
+public enum SNMPV3AuthenticationProvider
+{
+    MD5,
+    SHA1
+}
