+++ Source/NETworkManager.Models/Network/SNMPV3PrivacyProvider.cs
@@ -1,0 +1,10 @@
+﻿namespace NETworkManager.Models.Network;
+
+/// <summary>
+/// Enum for the SNMP v3 privacy provider.
+/// </summary>
+public enum SNMPV3PrivacyProvider
+{
+    DES,
+    AES
+}
