+++ Source/NETworkManager.Models/Network/SNMPV3Security.cs
@@ -1,0 +1,11 @@
+﻿namespace NETworkManager.Models.Network;
+
+/// <summary>
+/// Enum for the SNMP v3 security.
+/// </summary>
+public enum SNMPV3Security
+{
+    NoAuthNoPriv,
+    AuthNoPriv,
+    AuthPriv
+}
