+++ Source/NETworkManager.Models/Network/SNMPVersion.cs
@@ -1,0 +1,11 @@
+﻿namespace NETworkManager.Models.Network;
+
+/// <summary>
+/// Enum for the SNMP version.
+/// </summary>
+public enum SNMPVersion
+{
+    V1,
+    V2C,
+    V3
+}
